pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "interview-audio-analysis"
version = "0.1.0"
description = "Audio transcription and sentiment analysis pipeline"
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0",
    "pyannote.audio>=3.0",
    "huggingface_hub>=0.22",
    "pydub>=0.25",
    "openai-whisper>=20231117",
    "numpy>=1.24",
    "spacy>=3.0",
    "spacytextblob>=4.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.10",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
filterwarnings = [
    "ignore::DeprecationWarning:pydantic.*",
    "ignore::UserWarning:pyannote.*",
    "ignore::RuntimeWarning:pydub.*",
]
